+++ tbx/__init__.py
@@ -1,5 +1,8 @@
 """
 Toolbox
+
+This is free and unencumbered software released into the public domain.
+For more information, please refer to <http://unlicense.org/>
 """
 import contextlib
 import os