+++ tbx/verinfo.py
@@ -1,3 +1,8 @@
-# This file is named verinfo.py (rather than version.py) so that its name does
-# not conflict at import with the tbx method version()
-_v = '1.1.6'
+"""
+This file is named verinfo.py (rather than version.py) so that its name does
+not conflict at import with the tbx method version()
+
+This is free and unencumbered software released into the public domain.
+For more information, please refer to <http://unlicense.org/>
+"""
+_v = '1.1.7'