+++ test/conftest.py
@@ -1,5 +1,8 @@
 """
 Configure pytest
+
+This is free and unencumbered software released into the public domain.
+For more information, please refer to <http://unlicense.org/>
 """
 import os
 import pdb